requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
bottle==0.12.18
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
conan==1.25.1
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.0.7
distlib==0.3.0
distro==1.1.0
Django==3.0.5
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-js-asset==1.2.2
django-modeltranslation==0.14.4
djangorestframework==3.11.0
entrypoints==0.3
fasteners==0.15
filelock==3.0.12
future==0.18.2
idna==2.8
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.6
node-semver==0.6.1
notebook==6.0.3
packaging==20.0
pandocfilters==1.4.2
parso==0.7.0
patch-ng==1.17.4
pickleshare==0.7.5
Pillow==7.1.1
pluginbase==0.7
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
requests==2.22.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.17
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.11.2
